% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{arrange.parccvizieR_data}
\alias{arrange.parccvizieR_data}
\title{arrange dplyr wrapper}
\usage{
\method{arrange}{parccvizieR_data}(df, ...)
}
\arguments{
\item{df}{a tibble/data frame}

\item{...}{additional args}
}
\value{
data.frame a tibble/data frame with the same classes
}
\description{
arrange dplyr wrapper
}
