% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parccvizieR.R
\name{basic_read_and_clean}
\alias{basic_read_and_clean}
\title{Common/universal reading and cleaning of a data file}
\usage{
basic_read_and_clean(path)
}
\arguments{
\item{path}{full path to file, output of read_raw_results}
}
\value{
data.frame/tbl
}
\description{
Common/universal reading and cleaning of a data file
}
