% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{basic_summary_stats}
\alias{basic_summary_stats}
\title{Calculate basic summary stats.  Useful for initial 'how did we do' cuts and questions.}
\usage{
basic_summary_stats(pv_obj, studentids, subject_area = NA,
  academic_year = NA, grade_level_when_assessed = NA)
}
\arguments{
\item{pv_obj}{a conforming parccvizieR object}

\item{studentids}{vector of state_student_identifiers}

\item{subject_area}{optional, ELA or Math}

\item{academic_year}{optional, ending YYYY}

\item{grade_level_when_assessed}{optional, gr 3-11}
}
\value{
summary data frame
}
\description{
Calculate basic summary stats.  Useful for initial 'how did we do' cuts and questions.
}
