% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{build_growth_df}
\alias{build_growth_df}
\title{Build a growth df, given a start and end year}
\usage{
build_growth_df(srf, start, end, verbose = TRUE)
}
\arguments{
\item{srf}{a PARCC summative record file}

\item{start}{start academic year}

\item{end}{end academic year}

\item{verbose}{print status updates as the data is processed?
default is TRUE}
}
\value{
growth data frame
}
\description{
Build a growth df, given a start and end year
}
