% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parccvizieR.R
\name{clean_nj_2016_plus}
\alias{clean_nj_2016_plus}
\title{Clean columns present in 2016 SRFs and on, but not in the weird 2015 file}
\usage{
clean_nj_2016_plus(df)
}
\arguments{
\item{df}{data frame, product of basic_read}
}
\value{
data.frame/tbl
}
\description{
Clean columns present in 2016 SRFs and on, but not in the weird 2015 file
}
