% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cohort_year}
\alias{cohort_year}
\title{Calculate a student's cohort}
\usage{
cohort_year(end_year, grade)
}
\arguments{
\item{end_year}{4-digit year, where year is the ending year of
an academic year (eg 2015 is 2014-2015 school year).}

\item{grade}{student's grade level}
}
\value{
numeric vector
}
\description{
A student's cohort is defined as the year they would
graduate from 12th grade, assuming a normal pattern of promotion and
on time graduation.  A student can move between cohorts if they
are retained
}
