% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedupe.R
\name{dedupe_srf}
\alias{dedupe_srf}
\title{Dedupe a SRF to return one row per studentid per testcode per year}
\usage{
dedupe_srf(srf)
}
\arguments{
\item{srf}{a parcvizieR summative record file}
}
\value{
SRF with a rn column where 1 = best test record to use
}
\description{
Dedupe a SRF to return one row per studentid per testcode per year
}
