% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parccvizieR.R
\name{detect_result_file_layout}
\alias{detect_result_file_layout}
\title{Detect Layout/Origin of Raw PARCC Results File}
\usage{
detect_result_file_layout(df_path)
}
\arguments{
\item{df_path}{}
}
\value{
character string representing type detected
}
\description{
PARCC is a consortium, and (we believe) each state has some
flexibility in how it reports the core PARCC results.  For instance, MA has
a hybrid assessment that includes PARCC content and state content.  This
function detects the type of file by looking at the names/columns, and
processes accordingly.
}
