% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{generate_growth_data}
\alias{generate_growth_data}
\title{Generate growth data}
\usage{
generate_growth_data(srf)
}
\arguments{
\item{a}{PARCC summative record file.  assumes one row per student
per test code per year (eg output of dedupe_srf)}
}
\value{
data frame with multiple years of growth data
}
\description{
Generate growth data
}
