% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{next_test}
\alias{next_test}
\title{Returns sequential test code (next or prior)}
\usage{
next_test(x)
}
\arguments{
\item{x}{a test code}
}
\value{
the next or prior test code using standard PARCC progression
}
\description{
Returns sequential test code (next or prior)
}
