% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parccvizieR.R
\name{parccvizieR}
\alias{parccvizieR}
\title{Create a parccvizieR object}
\usage{
parccvizieR(results, local_roster = NA, verbose = TRUE)
}
\arguments{
\item{local_roster}{optional.  additional roster data to supplement
the data in the PARCC file.}

\item{verbose}{print status updates as the data is processed?
default is TRUE}

\item{results_location}{a file, or directory of files, containing PARCC
results.  parccvizieR supports NJ summative record files, but is
architected to gracefully extend support to other PARCC consortium states}
}
\description{
\code{parccvizieR} is the primary workflow function that generates the
parccvizieR object
}
\examples{
\dontrun{
pviz <- parccvizieR(sf = ex_sf)

is.parccvizieR(pviz)
}
}
