% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{parccvizieR_data}
\alias{parccvizieR_data}
\title{makes data a parccvizieR_data object}
\usage{
parccvizieR_data(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
same data frame with parccvizieR_data class
}
\description{
see https://github.com/tidyverse/dplyr/issues/719
}
