% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parccvizieR.R
\name{print.parccvizieR}
\alias{print.parccvizieR}
\title{Title print method for\code{parccvizier} class}
\usage{
\method{print}{parccvizieR}(x, ...)
}
\arguments{
\item{x}{a \code{parccvizier} object}

\item{...}{additional arguments}
}
\value{
some details about the object to the console
}
\description{
prints to console
}
