% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parccvizieR.R
\name{process_nj_srf15}
\alias{process_nj_srf15}
\alias{process_nj_srf16}
\alias{process_nj_srf17}
\alias{process_nj_srf18}
\title{Process a NJ SRF file (any year)}
\usage{
process_nj_srf15(path)

process_nj_srf16(path)

process_nj_srf17(path)

process_nj_srf18(path)
}
\arguments{
\item{path}{full path to file, output of read_raw_results}
}
\value{
data.frame/tbl
}
\description{
Process a NJ SRF file (any year)
}
