% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pv_limit_srf}
\alias{pv_limit_srf}
\title{Limit a parccvizieR SRF data frame}
\usage{
pv_limit_srf(pv_obj, studentids, subject_area = NA, academic_year = NA)
}
\arguments{
\item{pv_obj}{a conforming parccvizieR object}

\item{studentids}{vector of state_student_identifiers}

\item{subject_area}{optional, ELA or Math}

\item{academic_year}{optional, ending YYYY}
}
\description{
Limit a parccvizieR SRF data frame
}
