% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parccvizieR.R
\name{read_raw_results}
\alias{read_raw_results}
\title{Title}
\usage{
read_raw_results(x)
}
\arguments{
\item{x}{the results argument from parccvizieR.  either a path to one file
OR a folder of files}
}
\value{
character vector of full paths to results files
}
\description{
Title
}
