% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parccvizieR.R
\name{subject_area}
\alias{subject_area}
\title{Determine Test Subject Area}
\usage{
subject_area(x)
}
\arguments{
\item{x}{a vector of subjects}
}
\value{
vector of subject areas
}
\description{
Determine Test Subject Area
}
