% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary.parccvizieR_srf}
\alias{summary.parccvizieR_srf}
\title{Calculate basic summary stats using summary method.
Useful for initial 'how did we do' cuts and questions.}
\usage{
\method{summary}{parccvizieR_srf}(object, ...)
}
\arguments{
\item{object}{a \code{parccvizieR_srf} object}

\item{...}{other arguments to be passed to other functions (not currently supported)}
}
\value{
summary stats as a \code{parccvizieR_srf_summary} object.
}
