% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{ungroup.parccvizieR_data}
\alias{ungroup.parccvizieR_data}
\title{ungroup dplyr wrapper}
\usage{
\method{ungroup}{parccvizieR_data}(df, ...)
}
\arguments{
\item{df}{a tibble/data frame}

\item{...}{additional args}
}
\value{
data.frame a tibble/data frame with the same classes
}
\description{
ungroup dplyr wrapper
}
