% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parccvizieR.R
\name{parccvizieR}
\alias{parccvizieR}
\title{Create a parccvizieR object}
\usage{
parccvizieR(results, local_roster = NA, verbose = FALSE, ...)
}
\arguments{
\item{results}{a PARCC roster, or summative file}

\item{local_roster}{optional.  additional roster data to supplement
the data in the PARCC file.}

\item{...}{additional arguments}
}
\description{
\code{parccvizieR} is the primary workflow function that generates the
parccvizieR object
}
\examples{
\dontrun{
pviz <- parccvizieR(sf = ex_sf)

is.parccvizieR(pviz)
}
}
